pyproject.toml
[tool.poetry]
name = "goit-pyweb-11"
version = "0.1.0"
description = ""
authors = ["DeLight <nazarchukyayko@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
requests = "^2.32.3"
sqlalchemy = "^2.0.36"
psycopg2 = "^2.9.10"
asyncpg = "^0.30.0"
pydantic = "^2.10.4"
alembic = "^1.14.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.20"
fastapi-jwt-auth = "^0.5.0"
python-dotenv = "^1.0.1"
fastapi-mail = "^1.4.2"
redis = "^5.2.1"
fastapi-limiter = "^0.1.6"
cloudinary = "^1.42.1"
libgravatar = "^1.0.4"
passlib = "^1.7.4"
bcrypt = "^4.2.1"


[tool.poetry.group.dev.dependencies]
sphinx = "^8.1.3"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.4"
httpx = "^0.28.1"
pytest-asyncio = "^0.25.3"
aiosqlite = "^0.21.0"
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]